requirements.txt
streamlit
langchain
faiss-cpu
google-generativeai
sentence-transformers
PyMuPDF
python-dotenv
